Move shared examples in their own file
The idea is to split the gory details out of the main test file. This
will also run CI on each supported Ruby version and add more hints
about the fix in the previous commit working as expected :wink:.
